backend/requirements.txt
aerich==0.9.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dictdiffer==0.9.0
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
identify==2.6.10
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
pydantic==2.11.4
pydantic_core==2.33.2
pypika-tortoise==0.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.13.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
websockets==15.0.1
